pyproject.toml
[tool.poetry]
name = "drake-demos"
version = "0.1.0"
description = ""
authors = ["Griswald Brooks <griswald.brooks@gmail.com>"]
readme = "README.md"
packages = [{include = "drake_demos"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.2"
streamlit = "^1.25.0"
plotly = "^5.16.1"
drake = "^1.19.0"
matplotlib = "^3.7.2"
pandas = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cartpole = "drake_demos.cartpole_mpc:main"
